Adjust JS
