+++ q2_demux/_summarize/assets/src/main.js
@@ -8,10 +8,10 @@
 
 import initializePlot from './plot';
 
-export const init = (fwdScores, revScores=undefined) => { // eslint-disable-line import/prefer-default-export
-  const qualScores = { forward: Object.keys(fwdScores).map(key => [+key + 1, fwdScores[key]])}
+export const init = (fwdScores, revScores = undefined) => { // eslint-disable-line
+  const qualScores = { forward: Object.keys(fwdScores).map(key => [+key + 1, fwdScores[key]]) };
   if (revScores) {
-    qualScores['reverse'] = Object.keys(revScores).map(key => [+key + 1, revScores[key]])
+    qualScores.reverse = Object.keys(revScores).map(key => [+key + 1, revScores[key]]);
   }
   initializePlot(qualScores);
 };
